requirements.txt
python-dotenv>=1.0.0
pymongo>=4.0
requests>=2.28
tenacity>=8.0